apps/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
SQLAlchemy==2.0.36
asyncpg==0.29.0
alembic==1.13.3
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
httpx==0.27.2
sse-starlette==2.1.3
redis==5.0.8
orjson==3.10.7
